backend/requirements.txt
Django==4.2.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow>=9.0.0
daphne==4.0.0
channels==4.0.0
python-dateutil>=2.8.2
